pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-md-compiler"
version = "0.1.0"
description = "Compile CLAUDE.md into enforceable repo policy for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
cldc = "cldc.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
